[Silver III] Title: 주유소, Time: 368 ms, Memory: 35056 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -27,8 +27,8 @@ public class Main {
             city[i] = Integer.parseInt(st.nextToken());
         }
 
-        int min = city[0];
-        int answer = 0;
+        long min = city[0];
+        long answer = 0;
 
         for(int i=0;i<N-1;i++){
             if(min>city[i]) min = city[i];
